refactor: extract prompt history types to separate module
